refactor: move terratorch_iterate under src/claimed/iterate
Relocate the bundled iterate sources from the top-level terratorch_iterate/
package into src/claimed/iterate/, so all first-party packages live under a
single src/ root and the iterate code is namespaced under claimed.

- git mv terratorch_iterate -> src/claimed/iterate
- rewrite the import path terratorch_iterate -> claimed.iterate across all
  Python sources and tests, and in the coordinator plugin loader's built-in
  module list
- console scripts: iterate / iterate-classic now point at
  claimed.iterate.iterate2:main / claimed.iterate.main:main
- packages.find: drop the obsolete top-level terratorch_iterate* include;
  claimed.iterate is picked up by the existing claimed* rule under src/
- CI: flake8/pylint/docs path globs no longer reference terratorch_iterate/
- docs: mkdocstrings ':::terratorch_iterate.*' references -> ':::claimed.iterate.*'

No behavioral change. Verified: all iterate submodules import, coordinator
plugins load, iterate/iterate-classic entry points run, and unit +
coordinator integration tests pass (test_iterate2 shell-out included).
