+++ src/claimed/iterate/iterate2/plugin/coordinator/sqlite.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 import logging
 
-from terratorch_iterate.iterate2.plugin.coordinator import CoordinatorPlugin, register
+from claimed.iterate.iterate2.plugin.coordinator import CoordinatorPlugin, register
 
 logger = logging.getLogger("iterate2.coordinator.sqlite")
