--- terratorch_iterate/iterate2/plugin/__init__.py
@@ -1,1 +1,0 @@
-# terratorch_iterate.iterate2.plugin package
