+++ tests/integration/test_coordinator_plugins.py
@@ -33,7 +33,7 @@ def _fresh_registry():
     """Return a clean load_builtin_plugins / resolve_storage pair backed by an
     isolated registry so tests cannot leak state into each other."""
     # Re-import the coordinator package with a private registry copy.
-    from terratorch_iterate.iterate2.plugin import coordinator as coord_pkg
+    from claimed.iterate.iterate2.plugin import coordinator as coord_pkg
     import importlib, types
 
     # Build a fresh module clone with its own empty registry.
@@ -76,36 +76,36 @@ def _create_and_verify_study(storage: Any, study_name: str) -> None:
 
 class TestSQLiteCoordinator:
     def _make_storage(self, db_url: str):
-        from terratorch_iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
         return SQLiteCoordinator().get_storage(db_url)
 
     def test_matches_sqlite_scheme(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
         p = SQLiteCoordinator()
         assert p.matches("sqlite:///foo.db")
 
     def test_matches_dot_db_extension(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
         p = SQLiteCoordinator()
         assert p.matches("/tmp/my_study.db")
 
     def test_matches_dot_sqlite_extension(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
         p = SQLiteCoordinator()
         assert p.matches("/tmp/my_study.sqlite")
 
     def test_no_match_journalfs(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
         p = SQLiteCoordinator()
         assert not p.matches("js:///tmp/journal.log")
 
     def test_normalises_plain_path_to_sqlite_url(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
         url = SQLiteCoordinator().get_storage("/tmp/study.db")
         assert url.startswith("sqlite:///")
 
     def test_passthrough_existing_sqlite_url(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.sqlite import SQLiteCoordinator
         url = "sqlite:///existing.db"
         assert SQLiteCoordinator().get_storage(url) == url
 
@@ -116,7 +116,7 @@ class TestSQLiteCoordinator:
 
     def test_resolve_storage_via_registry(self, tmp_path):
         """End-to-end: resolve_storage() picks the SQLite plugin."""
-        from terratorch_iterate.iterate2.plugin.coordinator import (
+        from claimed.iterate.iterate2.plugin.coordinator import (
             load_builtin_plugins, resolve_storage,
         )
         load_builtin_plugins()
@@ -131,19 +131,19 @@ class TestSQLiteCoordinator:
 
 class TestJournalFSCoordinator:
     def _make_storage(self, journal_path: str):
-        from terratorch_iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
         return JournalFSCoordinator().get_storage(f"js:///{journal_path}")
 
     def test_matches_js_prefix(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
         assert JournalFSCoordinator().matches("js:///tmp/j.log")
 
     def test_no_match_sqlite(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
         assert not JournalFSCoordinator().matches("sqlite:///foo.db")
 
     def test_no_match_postgresql(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.journalfs import JournalFSCoordinator
         assert not JournalFSCoordinator().matches("postgresql://u:p@h/db")
 
     def test_returns_journal_storage_object(self, tmp_path):
@@ -184,7 +184,7 @@ class TestJournalFSCoordinator:
         assert not errors, f"Concurrent writers raised: {errors}"
 
     def test_resolve_storage_via_registry(self, tmp_path):
-        from terratorch_iterate.iterate2.plugin.coordinator import (
+        from claimed.iterate.iterate2.plugin.coordinator import (
             load_builtin_plugins, resolve_storage,
         )
         load_builtin_plugins()
@@ -211,24 +211,24 @@ postgres_required = pytest.mark.skipif(
 
 class TestPostgreSQLCoordinator:
     def test_matches_postgresql_scheme(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
         assert PostgreSQLCoordinator().matches("postgresql://u:p@h/db")
 
     def test_matches_legacy_postgres_scheme(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
         assert PostgreSQLCoordinator().matches("postgres://u:p@h/db")
 
     def test_no_match_sqlite(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
         assert not PostgreSQLCoordinator().matches("sqlite:///foo.db")
 
     def test_no_match_journalfs(self):
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
         assert not PostgreSQLCoordinator().matches("js:///foo.log")
 
     def test_normalises_legacy_scheme(self):
         """'postgres://' must be normalised to 'postgresql://' before it reaches SQLAlchemy."""
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import _extract_host
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import _extract_host
         # Test the helper that parses the host out of the normalised URL.
         host = _extract_host("postgresql://user:pass@my-host.example.com:5432/db")
         assert host == "my-host.example.com"
@@ -248,14 +248,14 @@ class TestPostgreSQLCoordinator:
             return real_import(name, *args, **kwargs)
 
         monkeypatch.setattr(builtins, "__import__", mock_import)
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
         with pytest.raises(ImportError, match="psycopg2"):
             PostgreSQLCoordinator().get_storage("postgresql://u:p@h/db")
 
     @postgres_required
     def test_full_study_lifecycle(self):
         import uuid
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
         storage = PostgreSQLCoordinator().get_storage(POSTGRES_URL)
         study_name = f"pg_integration_{uuid.uuid4().hex[:8]}"
         try:
@@ -270,7 +270,7 @@ class TestPostgreSQLCoordinator:
     @postgres_required
     def test_resolve_storage_via_registry(self):
         from optuna.storages import RDBStorage
-        from terratorch_iterate.iterate2.plugin.coordinator import (
+        from claimed.iterate.iterate2.plugin.coordinator import (
             load_builtin_plugins, resolve_storage,
         )
         load_builtin_plugins()
@@ -281,7 +281,7 @@ class TestPostgreSQLCoordinator:
     def test_parallel_trials(self):
         """Multiple threads sharing a PostgreSQL study must all complete cleanly."""
         import threading, uuid
-        from terratorch_iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
+        from claimed.iterate.iterate2.plugin.coordinator.postgresql import PostgreSQLCoordinator
         storage = PostgreSQLCoordinator().get_storage(POSTGRES_URL)
         study_name = f"pg_parallel_{uuid.uuid4().hex[:8]}"
         study = optuna.create_study(study_name=study_name, storage=storage,
